test: migrate usage to bdd
